third_party/WebKit/LayoutTests/fast/forms/input-select-api-support.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<input type="email" id="email">
<input type="number" id="number">
<input type="date" id="date">
<input type="month" id="month">
<input type="week" id="week">
<input type="time" id="time">
<input type="datetime-local" id="datetime-local">
<input type="color" id="color">
<input type="file" id="file">
<script>
description('Test to show select() support for Email, Number, Date, Month, Week, Time, DateTime-Local, Color and File.');

shouldNotThrow('document.getElementById("email").select()');
shouldNotThrow('document.getElementById("number").select()');
shouldNotThrow('document.getElementById("date").select()');
shouldNotThrow('document.getElementById("month").select()');
shouldNotThrow('document.getElementById("week").select()');
shouldNotThrow('document.getElementById("time").select()');
shouldNotThrow('document.getElementById("datetime-local").select()');
shouldNotThrow('document.getElementById("color").select()');
shouldNotThrow('document.getElementById("file").select()');

</script>
</body>
</html>
